clean up semantics demo
